refactor(user): move user entity to domain/user directory
Restructure the codebase by moving UserEntity and its tests from domain/entities to domain/user/entities for better organization and maintainability. Update all import paths accordingly.
