+++ src/config/__tests__/__helpers__/auth-datasource.helpers.ts
@@ -2,7 +2,7 @@ import { RegisterDto } from "@/domain/auth/dtos/register.dto";
 import { LoginDto } from "@/domain/auth/dtos/login.dto";
 import { LogoutDto } from "@/domain/auth/dtos/logout.dto";
 import { UpdateUserDto } from "@/domain/user/dtos/update-user.dto";
-import { UserEntity } from "@/domain/entities/user.entity";
+import { UserEntity } from "@/domain/user/entities/user.entity";
 import { AuthUserEntity } from "@/domain/auth/entities/auth-user.entity";
 
 // Mock types will be defined in individual test files
