+++ src/domain/auth/interfaces/__test__/auth-user.interfaces.test.ts
@@ -2,7 +2,7 @@ import {
   IAuthUserEntityFromRaw,
   IAuthUserEntity,
 } from "../auth-user.interfaces";
-import { UserEntity } from "@/domain/entities/user.entity";
+import { UserEntity } from "@/domain/user/entities/user.entity";
 import { AuthUserEntity } from "@/domain/auth/entities/auth-user.entity";
 
 describe("Auth User Interfaces", () => {
