+++ src/domain/auth/repositories/auth.repository.ts
@@ -4,7 +4,7 @@ import { RegisterDto } from "@/domain/auth/dtos/register.dto";
 import { RequestResetPasswordEmailDto } from "@/domain/auth/dtos/request-reset-password-email.dto";
 import { UpdateUserDto } from "@/domain/user/dtos/update-user.dto";
 import { AuthUserEntity } from "@/domain/auth/entities/auth-user.entity";
-import { UserEntity } from "@/domain/entities/user.entity";
+import { UserEntity } from "@/domain/user/entities/user.entity";
 
 export abstract class AuthRepository {
   abstract register(dto: RegisterDto): Promise<UserEntity | AuthUserEntity>;
