+++ src/domain/auth/use-cases/__test__/login-user.test.ts
@@ -1,7 +1,7 @@
 import { LoginUser } from "../login-user";
 import { LoginDto } from "@/domain/auth/dtos/login.dto";
 import { AuthUserEntity } from "@/domain/auth/entities/auth-user.entity";
-import { UserEntity } from "@/domain/entities/user.entity";
+import { UserEntity } from "@/domain/user/entities/user.entity";
 import { MockAuthRepository } from "@/config/__tests__/mocks/auth-repository.mock";
 
 describe("LoginUser", () => {
