+++ src/domain/user/entities/user.entity.ts
@@ -1,5 +1,5 @@
-import { TRawJson } from "../interfaces/general.interfaces";
-import { IUserData } from "../interfaces/user.interfaces";
+import { TRawJson } from "@/domain/interfaces/general.interfaces";
+import { IUserData } from "@/domain/interfaces/user.interfaces";
 import { ERRORS } from "@/config/strings/global.strings.json";
 
 export class UserEntity implements IUserData {
