+++ src/domain/user/use-cases/__test__/update-user.test.ts
@@ -1,6 +1,6 @@
 import { UpdateUser } from "@/domain/user/use-cases/update-user";
 import { UpdateUserDto } from "@/domain/user/dtos/update-user.dto";
-import { UserEntity } from "@/domain/entities/user.entity";
+import { UserEntity } from "@/domain/user/entities/user.entity";
 import { MockAuthRepository } from "@/config/__tests__/mocks/auth-repository.mock";
 
 describe("UpdateUser", () => {
