+++ src/infrastructure/datasources/auth.datasource.ts
@@ -1,6 +1,6 @@
 import { RegisterDto } from "@/domain/auth/dtos/register.dto";
 import { AuthUserEntity } from "@/domain/auth/entities/auth-user.entity";
-import { UserEntity } from "@/domain/entities/user.entity";
+import { UserEntity } from "@/domain/user/entities/user.entity";
 import { AuthRepository } from "@/domain/auth/repositories/auth.repository";
 import { AuthClient } from "../config/auth.client";
 import { ERRORS } from "@/config/strings/global.strings.json";
